+++ CHANGELOG.md
@@ -42,6 +42,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - fixed multiline
 - fixed piechart chunk selection syncing
 - fixed dashSelect
+- refactoring a multiline
 - fixed an empty data source will load indefinitely
 
 ## [1.84.0]